pyproject.toml
[project]
name = "agent-360-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.0",
    "gunicorn>=21.0",
]
